general optimization of code

+++ src/main/java/com/musalasoftdroneservice/repository/DroneRepository.java
@@ -1,7 +1,6 @@
 package com.musalasoftdroneservice.repository;
 
 import com.musalasoftdroneservice.entity.Drone;
-import com.musalasoftdroneservice.entity.Medication;
 import com.musalasoftdroneservice.enums.DroneState;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -16,7 +15,7 @@ public interface DroneRepository extends JpaRepository<Drone, Long> {
 
     Optional<Drone> findBySerialNumber(String serialNumber);
 
-    Optional<List<Drone>> findAllByDroneStateAndBatteryCapacityGreaterThan(DroneState droneState, BigDecimal batteryPercentage);
+    Optional<List<Drone>> findAllByDroneStateAndBatteryCapacityGreaterThan(DroneState droneState, int batteryPercentage);
 
     @Query(value = "SELECT battery_Capacity FROM drones WHERE serial_Number = ?1", nativeQuery = true)
     Optional<BigDecimal> getBatteryLevel(String serialNumber);
